Rename and consolidate "DelegateEndpoint" types
