+++ src/Http/Routing/src/Builder/RouteHandlerBuilder.cs
@@ -9,26 +9,26 @@ namespace Microsoft.AspNetCore.Builder
     /// <summary>
     /// Builds conventions that will be used for customization of MapAction <see cref="EndpointBuilder"/> instances.
     /// </summary>
-    public sealed class DelegateEndpointConventionBuilder : IEndpointConventionBuilder
+    public sealed class RouteHandlerBuilder : IEndpointConventionBuilder
     {
         private readonly IEnumerable<IEndpointConventionBuilder> _endpointConventionBuilders;
 
         /// <summary>
-        /// Instantiates a new <see cref="DelegateEndpointConventionBuilder" /> given a single
+        /// Instantiates a new <see cref="RouteHandlerBuilder" /> given a single
         /// <see cref="IEndpointConventionBuilder" />.
         /// </summary>
         /// <param name="endpointConventionBuilder">The <see cref="IEndpointConventionBuilder" /> to instantiate with.</param>
-        internal DelegateEndpointConventionBuilder(IEndpointConventionBuilder endpointConventionBuilder)
+        internal RouteHandlerBuilder(IEndpointConventionBuilder endpointConventionBuilder)
         {
             _endpointConventionBuilders = new List<IEndpointConventionBuilder>() { endpointConventionBuilder };
         }
 
         /// <summary>
-        /// Instantiates a new <see cref="DelegateEndpointConventionBuilder" /> given multiple
+        /// Instantiates a new <see cref="RouteHandlerBuilder" /> given multiple
         /// <see cref="IEndpointConventionBuilder" /> instances.
         /// </summary>
         /// <param name="endpointConventionBuilders">A sequence of <see cref="IEndpointConventionBuilder" /> instances.</param>
-        public DelegateEndpointConventionBuilder(IEnumerable<IEndpointConventionBuilder> endpointConventionBuilders)
+        public RouteHandlerBuilder(IEnumerable<IEndpointConventionBuilder> endpointConventionBuilders)
         {
             _endpointConventionBuilders = endpointConventionBuilders;
         }
